refactor: correct typo in filename
